+++ src/main/kotlin/com/github/srwi/pixellens/settings/PixelLensSettingsState.kt
@@ -8,6 +8,7 @@ import com.intellij.openapi.components.Storage
 @State(name = "PixelLensSettings", storages = [Storage("PixelLensSettings.xml")])
 class PixelLensSettingsState : PersistentStateComponent<PixelLensSettingsState> {
     var alwaysUseEvaluateTransmission: Boolean = false
+    var usePopupWindow: Boolean = false
 
     override fun getState(): PixelLensSettingsState {
         return this
@@ -15,6 +16,7 @@ class PixelLensSettingsState : PersistentStateComponent<PixelLensSettingsState> 
 
     override fun loadState(state: PixelLensSettingsState) {
         this.alwaysUseEvaluateTransmission = state.alwaysUseEvaluateTransmission
+        this.usePopupWindow = state.usePopupWindow
     }
 
     companion object {